+++ bee-signing/src/ternary/seed.rs
@@ -67,7 +67,7 @@ impl Seed {
     }
 
     /// Creates a new `Seed` from the current `Seed` and an index.
-    pub fn subseed(&self, index: u64) -> Self {
+    pub fn subseed(&self, index: usize) -> Self {
         let mut subseed = self.0.clone();
 
         for _ in 0..index {
